Use WeightedStackColorMapper and abstract base class
