+++ bokehjs/src/lib/models/annotations/color_bar.ts
@@ -3,7 +3,7 @@ import {Axis, CategoricalAxis, LinearAxis, LogAxis} from "../axes"
 import {TickFormatter} from "../formatters/tick_formatter"
 import {BasicTickFormatter, LogTickFormatter, CategoricalTickFormatter} from "../formatters"
 import {ColorMapper} from "../mappers/color_mapper"
-import {LinearColorMapper, LogColorMapper, ScanningColorMapper, CategoricalColorMapper, ContinuousColorMapper, StackColorMapper} from "../mappers"
+import {LinearColorMapper, LogColorMapper, ScanningColorMapper, CategoricalColorMapper, ContinuousColorMapper, WeightedStackColorMapper} from "../mappers"
 import {Range, Range1d, FactorRange} from "../ranges"
 import {Scale, LinearScale, LogScale, LinearInterpolationScale, CategoricalScale} from "../scales"
 import {Ticker} from "../tickers/ticker"
@@ -43,7 +43,7 @@ export class ColorBarView extends BaseColorBarView {
   get color_mapper(): ColorMapper {
     // Color mapper that is used to render this colorbar.
     let mapper = this.model.color_mapper
-    if (mapper instanceof StackColorMapper)
+    if (mapper instanceof WeightedStackColorMapper)
       mapper = mapper.alpha_mapper
     return mapper
   }
