+++ bokehjs/src/lib/models/mappers/index.ts
@@ -8,3 +8,4 @@ export {LogColorMapper}           from "./log_color_mapper"
 export {ScanningColorMapper}      from "./scanning_color_mapper"
 export {EqHistColorMapper}        from "./eqhist_color_mapper"
 export {StackColorMapper}         from "./stack_color_mapper"
+export {WeightedStackColorMapper} from "./weighted_stack_color_mapper"
